+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 
+### Changed
+- turning one shot off will stop the pad if it is playing
+
+
 ## v1.4.0 (2025-03-04)
 
 ### Changed